requirements.txt
opencv-python
sounddevice
numpy
openai-whisper
torch
tiktoken
requests
